+++ global.js
@@ -1,0 +1,5 @@
+const MSGTYPE_TEXT = 1;
+const SPECIAL_USERS = 'newsapp,fmessage,filehelper,weibo,qqmail,fmessage,tmessage,qmessage,qqsync,floatbottle,lbsapp,shakeapp,medianote,qqfriend,readerapp,blogapp,facebookapp,masssendapp,meishiapp,feedsapp,voip,blogappweixin,weixin,brandsessionholder,weixinreminder,officialaccounts,notification_messages,wxitil,userexperience_alarm,notification_messages';
+
+module.exports.MSGTYPE_TEXT = MSGTYPE_TEXT;
+module.exports.SPECIAL_USERS = SPECIAL_USERS;
